requirements.txt
# Core dependencies
paddlepaddle>=2.5.0
paddleocr>=2.7.0
Pillow>=9.0.0
numpy>=1.21.0
pandas>=1.3.0

# LLM and LoRA dependencies
torch>=2.0.0
transformers>=4.30.0
unsloth>=2024.1
vllm>=0.2.0
peft>=0.4.0

# Additional utilities
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.64.0

# Development and testing
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
